third_party/WebKit/LayoutTests/external/wpt/css/css-masking/parsing/mask-composite-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Masking Module Level 1: parsing mask-composite with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.fxtf.org/css-masking-1/#the-mask-composite">
<meta name="assert" content="mask-composite supports the full grammar '<compositing-operator>#' where <compositing-operator> is 'add | subtract | intersect | exclude'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("mask-composite", "add");
test_valid_value("mask-composite", "subtract");
test_valid_value("mask-composite", "intersect");
test_valid_value("mask-composite", "exclude");

test_valid_value("mask-composite", "intersect, intersect");
test_valid_value("mask-composite", "exclude, intersect, subtract, add");
</script>
</body>
</html>
